+++ lustre/mgs/mgs_handler.c
@@ -131,7 +131,7 @@ static int mgs_disconnect(struct obd_export *exp)
 
 static int mgs_handle(struct ptlrpc_request *req);
 
-static int mgs_completion_ast_config(struct ldlm_lock *lock, int flags,
+static int mgs_completion_ast_config(struct ldlm_lock *lock, __u64 flags,
                                      void *cbdata)
 {
         ENTRY;
@@ -151,7 +151,7 @@ static int mgs_completion_ast_config(struct ldlm_lock *lock, int flags,
         RETURN(ldlm_completion_ast(lock, flags, cbdata));
 }
 
-static int mgs_completion_ast_ir(struct ldlm_lock *lock, int flags,
+static int mgs_completion_ast_ir(struct ldlm_lock *lock, __u64 flags,
                                  void *cbdata)
 {
         ENTRY;
@@ -185,7 +185,7 @@ void mgs_revoke_lock(struct mgs_device *mgs, struct fs_db *fsdb, int type)
         ldlm_completion_callback cp = NULL;
         struct lustre_handle     lockh = { 0 };
         struct ldlm_res_id       res_id;
-        int flags = LDLM_FL_ATOMIC_CB;
+	__u64 flags = LDLM_FL_ATOMIC_CB;
         int rc;
         ENTRY;
